Refactor: Update .gitignore and move main activity files
This commit updates the `.gitignore` file to include the root `build/` directory and removes module-specific `.gitignore` files for `androidApp` and `shared` as they are now covered by the root `.gitignore`.

Additionally, files related to the main activity and application class have been moved to the root package of the `androidApp` module.

Other changes include:
- Removed example unit and instrumented tests.
- Deleted Room database schema JSON file.
- Removed downloadable fonts certificate file.
- Corrected `SystemBarStyle.auto` usage in `MainActivity`.
- Updated package names in `AndroidManifest.xml` due to file moves.